# frontend/openapi.json
{
  "components": {
    "headers": {
...
      "TraceId": {
        "description": "Request trace identifier.",
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "parameters": {
      "CsrfToken": {
        "description": "Token bound to the current opaque session.",
        "in": "header",
        "name": "X-CSRF-Token",
        "required": true,
        "schema": {
          "minLength": 1,
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Permission, assignment or CSRF validation failed.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
...
      "Problem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Request rejected.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
...
      "Unauthorized": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "A valid opaque session cookie is required.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      }
    },
    "schemas": {
...
      "Notification": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "message_key": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "read_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "kind",
          "message_key",
          "data",
          "read_at",
          "created_at"
        ],
        "type": "object"
      },
      "NotificationPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Notification"
            },
            "maxItems": 100,
            "type": "array"
          },
          "next_before_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "unread_count": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "items",
          "next_before_id",
          "unread_count"
        ],
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "trace_id"
        ],
        "type": "object"
      },
...
      "SuccessResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "description": "Opaque HttpOnly SameSite=Strict session cookie. Secure is mandatory outside loopback development.",
        "in": "cookie",
        "name": "dmx_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-host game server management API. Secret values are write-only.",
    "title": "DmxServerManager API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/notifications": {
      "get": {
        "operationId": "listNotifications",
        "parameters": [
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPage"
                }
              }
            },
            "description": "Current user's notification page."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/read-all": {
      "post": {
        "operationId": "readAllNotifications",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "All current-user notifications marked as read."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "put": {
        "operationId": "readNotification",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Notification marked as read."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}